+++ CHANGES.rst
@@ -6,6 +6,11 @@
 
 - Added support for log axes.
 
+- Fixed support for flipped limits (e.g. xmax < xmin)
+
+- Added support for setting vmin/vmax to functions to determine limits
+  on-the-fly.
+
 0.2 (2017-07-20)
 ----------------